Remove explicit control evaluator
